CO-1703 Move to shell module
