+++ src/index.ts
@@ -1,7 +1,7 @@
-import { spawn, exec } from "child_process";
 import { Application, Context, GitHubAPI } from "probot";
 import adapt from "probot-actions-adapter";
 
+import { shell, shellOutput, ShellError } from "./shell";
 import * as comment from "./comment";
 
 type UnwrapPromise<T> = T extends Promise<infer U> ? U : T;
@@ -133,72 +133,6 @@ const environmentIsAvailable = (context: Context, deployment?: Deployment) => {
   } else {
     return true;
   }
-};
-
-class ShellError extends Error {
-  constructor(public message: string, public output: string) {
-    super(message);
-  }
-}
-
-const shell = async (
-  commands: string[],
-  extraEnv: Record<string, string> = {}
-): Promise<string> => {
-  const output: string[] = [];
-  return new Promise((resolve, reject) => {
-    const env = {
-      ...process.env,
-      ...extraEnv,
-    };
-    const options = { env, cwd: process.cwd() };
-    // TODO shell escape command
-    const child = spawn(
-      "bash",
-      ["-e", "-x", "-c", commands.join("\n")],
-      options
-    );
-    child.stdout.on("data", (data) => {
-      const str = data.toString();
-      output.push(str);
-      console.log(str);
-    });
-    child.stderr.on("data", (data) => {
-      const str = data.toString();
-      output.push(str);
-      console.error(str);
-    });
-    child.on("error", (e) => {
-      reject(e);
-    });
-    child.on("exit", (code) => {
-      if (code === 0) {
-        resolve(output.join("\n"));
-      } else {
-        reject(
-          new ShellError(`exited with status code ${code}`, output.join("\n"))
-        );
-      }
-    });
-  });
-};
-
-const shellOutput = (command: string): Promise<string> => {
-  return new Promise((resolve, reject) => {
-    exec(command, (e, stdout, stderr) => {
-      if (stderr) {
-        console.error(stderr);
-      }
-      if (stdout) {
-        console.log(stdout);
-      }
-      if (e) {
-        reject(new ShellError(e.message, [stdout, stderr].join("\n")));
-      } else {
-        resolve(stdout);
-      }
-    });
-  });
 };
 
 const handleDeploy = async (
